### 2018-03-30 #658 @arfon
:wave: @momargoh - many thanks for your submission to JOSS. I noticed that your project is licensed using a Creative Commons license which is not recommended for software ([by Creative Commons](https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software)) and not allowable in JOSS.

Would you be able to swap the license for one of the [OSI-approved](https://opensource.org/licenses) ones?
