### 2018-03-31 2767641
Update license to OSI approved licence: Apache License 2.0
