+++ mcycle/__init__.py
@@ -2,7 +2,7 @@ r"""
 :Abstract: *mcycle* is a power cycle analysis/sizing package. All units are SI. The project homepage can be found at `<https://github.com/momargoh/mcycle>`_, the documentation can be found at `<https://mcycle.readthedocs.io>`_.
 :Author: Momar Hughes
 :Contact: momar.hughes@unsw.edu.au
-:License: MIT
+:License: Apache License 2.0
 :Requirements: numpy, scipy, matplotlib, CoolProp
 """
 from . import DEFAULTS
