### 2018-03-31 #658 @momargoh
Hi Arfon, pleasure to meet you and thanks for taking my paper into consideration!

OK, I completely overlooked that before, no worries, I've updated the licence to Apache License 2.0. I've also updated the package version to v0.1.3 because I need to update the licence on PyPI as well. https://github.com/momargoh/MCycle/commit/27676417c080d4662795d859d85e9391d942dca9
