### 2018-03-31 4019259
Correction to setup.py license classifier
