### 2018-04-24 #658 @katyhuff
@kyleniemeyer @trallard @nnadeau : Would any of you be willing to edit this submission? I think you would be very fitting. The topic is "MCycle: A Python package for 1D sizing and analysis of thermodynamic power cycles." It's a python package. From the paper:

> MCycle allows users to analyse thermodynamic power cycles and their individual com- ponents, as well as size cycle components to de ned cycle design parameters. It was developed due to a need for an open source tool with easy scripting for sizing heat ex- changers that would allow control over individual component parameters. Users may now analyse a growing collection of thermodynamic power cycles and cycle components, including heat exchangers, expanders, compressors, heaters and coolers. Each type of component has models of varying complexity, allowing MCycle to be equally applica- ble to simplistic cycle analyses as to detailed sizing optimisations. The project page is found at https://github.com/momargoh/MCycle and the documentation is hosted at https://mcycle.readthedocs.io.
>
> To evaluate a cycle’s  fluid properties, MCycle uses the Python wrapper of the open-source thermodynamic properties library CoolProp (Bell et al. 2014). A library of heat transfer and component analysis methods is provided , containing theoretical relations and semi- empirical correlations sourced from published research articles (refer to the documentation for specific references).
