### 2018-05-01 #710 @whedon
[REVIEW]: MCycle: A Python package for 1D sizing and analysis of thermodynamic power cycles
**Submitting author:** @momargoh (<a href="http://orcid.org/0000-0002-6928-2187">Momar Hughes</a>)
**Repository:** <a href="https://github.com/momargoh/MCycle" target ="_blank">https://github.com/momargoh/MCycle</a>
**Version:** v0.1.2
