### 2018-05-09 #710 @katyhuff
@Tabeel there's a little note about this up at the top of the issue, but it's easy to miss.

It says:

 If you cannot edit the checklist please:
