+++ mcycle/components/coolers/__init__.py
@@ -1,0 +1,4 @@
+"""Coolers"""
+#import pyximport
+#pyximport.install()
+from .clr_basic import ClrBasic, ClrBasicConstP, ClrBasicConstV
