+++ mcycle/geometries/__init__.py
@@ -1,0 +1,3 @@
+#import pyximport
+#pyximport.install()
+from .geom_hxplate import *
