+++ mcycle/components/heaters/htr_basic.pyx
@@ -1,0 +1,266 @@
+from ...bases.component cimport Component11
+from ...bases.config cimport Config
+from ...bases.flowstate cimport FlowState
+from ...bases.mcabstractbase cimport MCAttr
+import CoolProp as CP
+
+cdef class HtrBasic(Component11):
+    r"""Basic expansion defined by a pressure ratio and isentropic efficiency.
+
+Parameters
+----------
+pRatio : float
+    Pressure increase ratio [-].
+effIsentropic : float, optional
+    Isentropic efficiency [-]. Defaults to 1.
+flowIn : FlowState, optional
+    Incoming FlowState. Defaults to None.
+flowOut : FlowState, optional
+    Outgoing FlowState. Defaults to None.
+ambient : FlowState, optional
+    Ambient environment flow state. Defaults to None.
+sizeAttr : string, optional
+    Default attribute used by size(). Defaults to "pRatio".
+sizeBracket : float or list of float, optional
+    Bracket containing solution of size(). Defaults to [1, 50].
+
+    - if sizeBracket=[a,b]: scipy.optimize.brentq is used.
+
+    - if sizeBracket=a or [a]: scipy.optimize.newton is used.
+name : string, optional
+    Description of Component object. Defaults to "ExpBasic instance".
+notes : string, optional
+    Additional notes on the component such as model numbers. Defaults to "No notes/model info.".
+config : Config, optional
+    Configuration parameters. Defaults to the default Config object.
+kwargs : optional
+    Arbitrary keyword arguments.
+    """
+
+    def __init__(self,
+                 double Q,
+                 double effThermal=1.0,
+                 FlowState flowIn=None,
+                 FlowState flowOut=None,
+                 FlowState ambient=None,
+                 str sizeAttr="pRatio",
+                 list sizeBracket=[1, 50],
+                 list sizeUnitsBracket=[],
+                 str name="HtrBasic instance",
+                 str notes="No notes/model info.",
+                 Config config=Config()):
+        super().__init__(flowIn, flowOut, ambient, sizeAttr, sizeBracket, sizeUnitsBracket, [0, 0], name, notes,
+                         config)
+        assert (
+            effThermal > 0 and effThermal <= 1.
+        ), "Thermal efficiency={} is not in range (0, 1]".format(effThermal)
+        self.Q = Q
+        self.effThermal = effThermal
+        
+        self._inputs = {"Q": MCAttr(float, "power"), "effThermal": MCAttr(float, "none"),
+                        "flowIn": MCAttr(FlowState, "none"), "flowOut": MCAttr(FlowState, "none"), 'ambient': MCAttr(FlowState, 'none'), "sizeAttr": MCAttr(str, "none"),
+                "sizeBracket": MCAttr(list, "none"),"sizeUnitsBracket": MCAttr(list, "none"), "name": MCAttr(str, "none"), "notes": MCAttr(str, "none"),
+                        "config": MCAttr(Config, "none")}
+        self._properties= {"mWf": MCAttr(float, "mass/time"), "dpWf()": MCAttr(float, "pressure"),
+                "dpSf()": MCAttr(float, "pressure")}
+
+    cpdef public double _effFactorWf(self):
+        return 1
+    
+    cpdef public double _effFactorSf(self):
+        return self.effThermal
+
+    cpdef public double  dpWf(self):
+        """float: Pressure drop of the working fluid [Pa]. Defaults to 0."""
+        return 0
+
+    cpdef public double dpSf(self):
+        """float: Pressure drop of the secondary fluid [Pa]. Defaults to 0."""
+        return 0
+
+    cpdef public void run(self):
+        pass
+    
+
[…]

[…]
