### 2018-06-15 #710 @arfon
@puncan - it looks like you're only part way through your review here. Any chance you could complete this in the next week or so?
