### 2018-06-15 #710 @puncan
Yes, my apologies. It got away from me. Certainly within a week.

-Paul

On Fri, Jun 15, 2018 at 1:02 PM Arfon Smith <notifications@github.com>
wrote:

> @puncan <https://github.com/puncan> - it looks like you're only part way
> through your review here. Any chance you could complete this in the next
> week or so?
>
