### 2018-06-22 #710 @puncan
There were a couple I checked off after creating new issues. Am I doing this right?
