### 2018-06-22 #710 @momargoh
Thanks @puncan I'll fix up the issues you've raised now.
