### 2018-06-22 #710 @puncan
I don’t see why not.

Sent from my iPhone

> On Jun 21, 2018, at 7:20 PM, Momar Hughes <notifications@github.com> wrote:
>
> Also, there is a newer version of MCycle that uses Cython; it has a different build and different directory structure. At the moment it's just in the cython fork, I haven't pulled it to master yet. Could I update my joss submission to consider this version?
>
