### 2018-07-19 972704a
Major updates
Updated source files and documentation. Added quick start example. Added logging feature instead of printing warnings.
