+++ examples/hughes2017/hughes2017.py
@@ -1,293 +1,4 @@
-r"""Basic organic Rankine cycle with a corrugated-plate evaporator
-======================================================================
-
-.. contents::
-
-This example is found in examples/hughes2017 and is an update to the analysis found in the paper *On the potential of organic Rankine cycles for recovering waste heat from the exhaust gases of a piston aero-engine* by Hughes and Olsen, presented at Asia-Pacific Symposium on Aerospace Technology 2017, Seoul, Korea (`hyperlink <https://www.unsworks.unsw.edu.au/primo-explore/fulldisplay?docid=unsworks_modsunsworks_47631&context=L&vid=UNSWORKS&search_scope=unsworks_search_scope&tab=default_tab&lang=en_US>`_) [Hughes2017]_.
-
-The first section of the script imports the required modules and sets up formatting for the plots, including a folder to store the plots, the desired file extension (.png or .jpg) and some custom formatting for matplotlib.::
-
-    >>> import mcycle as mc
-    >>> import os
-    >>> import numpy as np
-    >>> import CoolProp as CP
-    >>> import matplotlib as mpl
-    >>> import matplotlib.pyplot as plt
-
-    >>> plots_folder = "plots"
-    >>> plots_format = "png"  # change to "jpg" if preferred
-    >>> plots_dpi = 1200
-    >>> if not os.path.exists(plots_folder):
-            os.makedirs(plots_folder)
-
-    >>> mpl.rc("lines", lw=2.0)
-    >>> mpl.rc("lines", markersize=2)
-    >>> colours = ["blue", "red", "green", "black"]
-    >>> linestyles = ["-", "--", "-.", ":"]
-    >>> markerstyles = ["o", "v", "s", "^"]
-
-Setting up the default characteristics of the cycle
--------------------------------------------------------------
-
-
-Working fluid
-**********************************************
-
-R123 (using CoolProp's HEOS library).::
-
-    >>> wf = mc.FlowState("R123", "HEOS", None)
-
-Heat source
-**********************************************
-
-The exhaust flow from the engine is approximated using the following characteristics.
-
-- :math:`N_2` (75%), :math:`C0_2` (25%) (by molar fraction)
-- :math:`p_{static} = 1.116` bar
-- :math:`T_{static} = 1170` K
-- :math:`\dot{m} = 0.09` Kg/s
-
-As CoolProp can have difficulties with mixtures, we instead use FlowStatePoly for the exhaust gasses; fitting a 2nd degree polynomial to the properties of the exhaust mixture evaluated at 300, 500, 700, 900 and 1170 K::
-
-    >>> TData = [300, 500, 700, 900, 1170.]
-    >>> sourceRefData = mc.RefData("N2[0.75]&CO2[0.25], 2, 1.116e5, TData, phase=CP.iphase_gas)
-    >>> sourceIn = mc.FlowStatePoly(sourceRefData, 0.09, CP.PT_INPUTS, 1.116e5, 1170.)
-
-Ambient air/ heat sink
-**********************************************
-
-Assuming the engine is travelling at 100knots at an altitude of 3000ft, the ambient air is approximated using the following characteristics.
-
-- Air (100%)
-- :math:`p_{static} = 0.8826` bar
-- :math:`T_{static} = 281.65` K
-
-::
-
-    >>> sinkIn = mc.FlowState("Air", "HEOS", None, None, CP.PT_INPUTS, 0.8826e5, 281.65)
-
-The dead state properties of the source and sink are at this same pressure and temperature.::
-
-    >>> sinkDead = sinkIn.copy()
-    >>> sourceDead = mc.FlowState(sourceFluid, "HEOS", CP.iphase_gas, None, CP.PT_INPUTS, 0.8826e5, 281.65)
-
-
-Evaporator and condenser
-**********************************************
-
[…]
