### 2018-07-23 #710 @katyhuff
Yes, go right ahead and update to the new version. Sorry for the delay in responding. It looks like all of the issues raised by the reviewers have now been handled. Is that correct @puncan @Tabeel @momargoh ?

If so, then @momargoh  at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
