### 2018-07-24 #710 @momargoh
No, I haven't handled all the issues raised by @puncan yet, doing so now. I'm having trouble committing the new Cython-based version to pip due to problems creating a cross-platform dist wheel. The package can be built from the source using `python3 setup.py install` and similarly the docs will not be hosted properly until the package is on pip, so at the moment they can just be built from the source using Sphinx3 `make html`.
Is this OK: the package needing to be built from source instead of available from pip?
