### 2018-07-24 200b06c
Numerous updates and debugging
Major changes are:

  - added quickstart example
  - updated the docs/ folder
  - sorted out bugs in RankineBasic and HxPlate
  - added initial logging functionality
  - updated README

Currently this version of the source code will not upload to pip due to problems making a cross-platform wheel. This needs to be sorted out. For now, the pip version is deprecated and mcycle must be installed from source.
