+++ mcycle/bases/cycle.c
@@ -1443,9 +1443,9 @@ extern int __pyx_module_is_main_mcycle__bases__cycle;
 int __pyx_module_is_main_mcycle__bases__cycle = 0;
 
 /* Implementation of 'mcycle.bases.cycle' */
+static PyObject *__pyx_builtin_print;
 static PyObject *__pyx_builtin_super;
 static PyObject *__pyx_builtin_AttributeError;
-static PyObject *__pyx_builtin_print;
 static const char __pyx_k__2[] = "";
 static const char __pyx_k__4[] = "()";
 static const char __pyx_k__5[] = "\n";
@@ -2229,8 +2229,8 @@ void __pyx_f_6mcycle_5bases_5cycle_5Cycle_update(struct __pyx_obj_6mcycle_5bases
  *         cdef str key
  *         cdef list key_split
  *         cdef dict store = {}             # <<<<<<<<<<<<<<
+ *         print(kwargs)
  *         for key, value in kwargs.items():
- *             if hasattr(self, 'set_{}'.format(key)):
  */
   __pyx_t_1 = __Pyx_PyDict_NewPresized(0); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 53, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
@@ -2240,6 +2240,23 @@ void __pyx_f_6mcycle_5bases_5cycle_5Cycle_update(struct __pyx_obj_6mcycle_5bases
   /* "mcycle/bases/cycle.pyx":54
  *         cdef list key_split
  *         cdef dict store = {}
+ *         print(kwargs)             # <<<<<<<<<<<<<<
+ *         for key, value in kwargs.items():
+ *             if hasattr(self, 'set_{}'.format(key)):
+ */
+  __pyx_t_1 = PyTuple_New(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 54, __pyx_L1_error)
+  __Pyx_GOTREF(__pyx_t_1);
+  __Pyx_INCREF(__pyx_v_kwargs);
+  __Pyx_GIVEREF(__pyx_v_kwargs);
+  PyTuple_SET_ITEM(__pyx_t_1, 0, __pyx_v_kwargs);
+  __pyx_t_2 = __Pyx_PyObject_Call(__pyx_builtin_print, __pyx_t_1, NULL); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 54, __pyx_L1_error)
+  __Pyx_GOTREF(__pyx_t_2);
+  __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
+  __Pyx_DECREF(__pyx_t_2); __pyx_t_2 = 0;
+
+  /* "mcycle/bases/cycle.pyx":55
+ *         cdef dict store = {}
+ *         print(kwargs)
  *         for key, value in kwargs.items():             # <<<<<<<<<<<<<<
  *             if hasattr(self, 'set_{}'.format(key)):
  *                 getattr(self, 'set_{}'.format(key))(value)
@@ -2247,90 +2264,90 @@ void __pyx_f_6mcycle_5bases_5cycle_5Cycle_update(struct __pyx_obj_6mcycle_5bases
   __pyx_t_6 = 0;
   if (unlikely(__pyx_v_kwargs == Py_None)) {
[…]
