+++ paper.bib
@@ -1,0 +1,12 @@
+@article{bell2014coolprop,
+  author    = {Ian H. Bell and Jorrit Wronski and Sylvain Quoilin and Vincent Lemort},
+  title     = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation and the Open-Source Thermophysical Property Library {CoolProp}},
+  journal   = {Industrial {\&} Engineering Chemistry Research},
+  year      = {2014},
+  volume    = {53},
+  number    = {6},
+  pages     = {2498--2508},
+  month     = {jan},
+  doi       = {10.1021/ie4033999},
+  publisher = {American Chemical Society ({ACS})},
+}
