### 2018-07-25 #710 @momargoh
OK, the docs are building OK https://mcycle.readthedocs.io/index.html. For now the source has to be installed from source as opposed to from pip. I have updated my JOSS submission to v1.0.0: the latest version of the source code.
