### 2018-07-25 fa1a035
minor update to docs readme link
