### 2018-07-25 1cd14b9
add Cython to install_requires
