### 2018-07-26 ee7d576
Updated runBracket to take absolute values
runBracket now requires absolute values for the enthalpy bounds instead of the previous attempt of bounds 'relative' to the min and crit temperatures. Significantly more robust even if it requires a little more user attention
