+++ mcycle/components/hxs/hx_plate.c
@@ -1811,10 +1811,6 @@ static int *__pyx_vp_6mcycle_8DEFAULTS_MAXITER_COMPONENT = 0;
 #define __pyx_v_6mcycle_8DEFAULTS_MAXITER_COMPONENT (*__pyx_vp_6mcycle_8DEFAULTS_MAXITER_COMPONENT)
 static int *__pyx_vp_6mcycle_8DEFAULTS_MAX_WALLS = 0;
 #define __pyx_v_6mcycle_8DEFAULTS_MAX_WALLS (*__pyx_vp_6mcycle_8DEFAULTS_MAX_WALLS)
-static double *__pyx_vp_6mcycle_8DEFAULTS_RUN_BRACKET_MIN_H = 0;
-#define __pyx_v_6mcycle_8DEFAULTS_RUN_BRACKET_MIN_H (*__pyx_vp_6mcycle_8DEFAULTS_RUN_BRACKET_MIN_H)
-static double *__pyx_vp_6mcycle_8DEFAULTS_RUN_BRACKET_MAX_H = 0;
-#define __pyx_v_6mcycle_8DEFAULTS_RUN_BRACKET_MAX_H (*__pyx_vp_6mcycle_8DEFAULTS_RUN_BRACKET_MAX_H)
 static double *__pyx_vp_6mcycle_8DEFAULTS_GRAVITY = 0;
 #define __pyx_v_6mcycle_8DEFAULTS_GRAVITY (*__pyx_vp_6mcycle_8DEFAULTS_GRAVITY)
 static PyObject **__pyx_vp_6mcycle_8DEFAULTS_COOLPROP_EOS = 0;
@@ -13058,7 +13054,7 @@ static PyTypeObject __pyx_type_6mcycle_10components_3hxs_8hx_plate_HxPlate = {
   0, /*tp_setattro*/
   0, /*tp_as_buffer*/
   Py_TPFLAGS_DEFAULT|Py_TPFLAGS_HAVE_VERSION_TAG|Py_TPFLAGS_CHECKTYPES|Py_TPFLAGS_HAVE_NEWBUFFER|Py_TPFLAGS_BASETYPE|Py_TPFLAGS_HAVE_GC, /*tp_flags*/
-  "HxPlate(unicode flowSense=u'counterflow', int NPlate=3, double RfWf=0, double RfSf=0, SolidMaterial plate=None, double tPlate=nan, Geom geomPlateWf=None, Geom geomPlateSf=None, double L=nan, double W=nan, double ARatioWf=1, double ARatioSf=1, double ARatioPlate=1, double DPortWf=nan, double DPortSf=nan, double LVertPortWf=nan, double LVertPortSf=nan, list coeffs_LPlate=[0, 1], list coeffs_WPlate=[0, 1], list coeffs_weight=[], double effThermal=1.0, FlowState flowInWf=None, FlowState flowInSf=None, FlowState flowOutWf=None, FlowState flowOutSf=None, FlowState ambient=None, unicode sizeAttr=u'NPlate', list sizeBracket=[3, 100], list sizeUnitsBracket=[1e-5, 10.], runBracket=[RUN_BRACKET_MIN_H, RUN_BRACKET_MAX_H], unicode name=u'HxBasic instance', unicode notes=u'No notes/model info.', Config config=Config(), _unitClass=HxUnitPlate)\nCharacterises a basic plate heat exchanger consisting of alternating working fluid and secondary fluid flows separated by a solid wall with single-phase or multi-phase working fluid but only single-phase secondary fluid.\n\nParameters\n----------\nflowSense : str, optional\n    Relative direction of the working and secondary flows. May be either \"counterflow\" or \"parallel\". Defaults to \"counterflow\".\nRfWf : floa
[…]
