### 2018-07-26 a9dbe36
Renamed bracket to bounds
