+++ tests/test_HtrBasic.py
@@ -27,7 +27,7 @@ class TestHtrBasic(unittest.TestCase):
         flowOut = mc.FlowState("He", -1, -1, CP.PT_INPUTS, 6.7e6, 1190)
 
         htr = mc.HtrBasicConstP(3.975e6, 1.0, flowIn, flowOut)
-        htr.update({'sizeAttr': 'm', 'sizeBracket': [0.8, 1.1]})
+        htr.update({'sizeAttr': 'm', 'sizeBounds': [0.8, 1.1]})
         htr.size()
         self.assertAlmostEqual(htr.m, 1.000, 3)
 
