+++ mcycle/bases/geom.c
@@ -1316,9 +1316,9 @@ static const char __pyx_k_summary_2[] = "summary";
 static const char __pyx_k_pyx_result[] = "__pyx_result";
 static const char __pyx_k_pyx_vtable[] = "__pyx_vtable__";
 static const char __pyx_k_rstHeading[] = "rstHeading";
+static const char __pyx_k_sizeBounds[] = "sizeBounds";
 static const char __pyx_k_validClass[] = "validClass";
 static const char __pyx_k_PickleError[] = "PickleError";
-static const char __pyx_k_sizeBracket[] = "sizeBracket";
 static const char __pyx_k_printSummary[] = "printSummary";
 static const char __pyx_k_propertyKeys[] = "propertyKeys";
 static const char __pyx_k_pyx_checksum[] = "__pyx_checksum";
@@ -1328,7 +1328,7 @@ static const char __pyx_k_Geom_instance[] = "Geom instance";
 static const char __pyx_k_reduce_cython[] = "__reduce_cython__";
 static const char __pyx_k_pyx_PickleError[] = "__pyx_PickleError";
 static const char __pyx_k_setstate_cython[] = "__setstate_cython__";
-static const char __pyx_k_sizeUnitsBracket[] = "sizeUnitsBracket";
+static const char __pyx_k_sizeUnitsBounds[] = "sizeUnitsBounds";
 static const char __pyx_k_mcycle_bases_geom[] = "mcycle.bases.geom";
 static const char __pyx_k_pyx_unpickle_Geom[] = "__pyx_unpickle_Geom";
 static const char __pyx_k_cline_in_traceback[] = "cline_in_traceback";
@@ -1383,8 +1383,8 @@ static PyObject *__pyx_n_s_rstHeading;
 static PyObject *__pyx_n_s_setstate;
 static PyObject *__pyx_n_s_setstate_cython;
 static PyObject *__pyx_n_u_sizeAttr;
-static PyObject *__pyx_n_u_sizeBracket;
-static PyObject *__pyx_n_u_sizeUnitsBracket;
+static PyObject *__pyx_n_u_sizeBounds;
+static PyObject *__pyx_n_u_sizeUnitsBounds;
 static PyObject *__pyx_kp_s_stringsource;
 static PyObject *__pyx_kp_u_summary;
 static PyObject *__pyx_n_s_summary_2;
@@ -2201,7 +2201,7 @@ static PyObject *__pyx_pf_6mcycle_5bases_4geom_4Geom_4summary(struct __pyx_obj_6
     /* "mcycle/bases/geom.pyx":60
  *             ]:
  *                 pass
- *             elif k in ["sizeAttr", "sizeBracket", "sizeUnitsBracket"]:             # <<<<<<<<<<<<<<
+ *             elif k in ["sizeAttr", "sizeBounds", "sizeUnitsBounds"]:             # <<<<<<<<<<<<<<
  *                 pass
  *             elif k in ["name", "notes", "config"]:
  */
@@ -2213,13 +2213,13 @@ static PyObject *__pyx_pf_6mcycle_5bases_4geom_4Geom_4summary(struct __pyx_obj_6
       __pyx_t_1 = __pyx_t_2;
       goto __pyx_L15_bool_binop_done;
     }
[…]
