### 2018-07-29 #710 @katyhuff
It looks to me like @puncan approved this a few comments ago. (https://github.com/openjournals/joss-reviews/issues/710#issuecomment-407145425)

It seems you can go ahead and make the DOI whenever you're ready.
