### 2018-07-31 742c6c5
Minor edits to paper.md and conversions
