+++ mcycle/library/conversions.h
@@ -1,0 +1,34 @@
+/* Generated by Cython 0.27.3 */
+
+#ifndef __PYX_HAVE__mcycle__library__conversions
+#define __PYX_HAVE__mcycle__library__conversions
+
+
+#ifndef __PYX_HAVE_API__mcycle__library__conversions
+
+#ifndef __PYX_EXTERN_C
+  #ifdef __cplusplus
+    #define __PYX_EXTERN_C extern "C"
+  #else
+    #define __PYX_EXTERN_C extern
+  #endif
+#endif
+
+#ifndef DL_IMPORT
+  #define DL_IMPORT(_T) _T
+#endif
+
+__PYX_EXTERN_C double Pa2bar(double, int __pyx_skip_dispatch);
+
+#endif /* !__PYX_HAVE_API__mcycle__library__conversions */
+
+/* WARNING: the interface of the module init function changed in CPython 3.5. */
+/* It now returns a PyModuleDef instance instead of a PyModule instance. */
+
+#if PY_MAJOR_VERSION < 3
+PyMODINIT_FUNC initconversions(void);
+#else
+PyMODINIT_FUNC PyInit_conversions(void);
+#endif
+
+#endif /* !__PYX_HAVE__mcycle__library__conversions */
