### 2018-07-31 fdd01dc
Temp remove Zarati example
