+++ examples/zarati2017/zarati2017.py
@@ -1,157 +1,0 @@
-import mcycle as mc
-import os
-import numpy as np
-from math import nan
-import CoolProp as CP
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-
-
-@mc.timeThis
-def run_example():
-    """run example: zarati2017.py"""
-    # ************************************************
-    # Set up formatting
-    # ************************************************
-    # cycle.plot parameters
-    mc.DEFAULTS.PLOT_DIR = 'plots'  # directory to save plots into
-    mc.DEFAULTS.PLOT_FORMAT = 'png'  # change to "jpg" if preferred
-    mc.DEFAULTS.PLOT_DPI = 600  # dpi (resolution) of plots
-    mc.DEFAULTS.updateDefaults()  # check the defaults are OK
-    # custom format for matplotlib
-    mpl.rc("lines", lw=2.0)
-    mpl.rc("lines", markersize=2)
-    colours = ["blue", "red", "green", "black"]
-    linestyles = ["-", "--", "-.", ":"]
-    markerstyles = ["o", "v", "s", "^"]
-
-    # ************************************************
-    # Set up default cycle parameters
-    # ************************************************
-    print("Begin setting up cycle...")
-    # Create basic compressor object
-    comp = mc.CompBasic(pRatio=nan, effIsentropic=0.9, sizeAttr="pRatio")
-    # Create basic expander object
-    exp = mc.ExpBasic(pRatio=nan, effIsentropic=0.85, sizeAttr="pRatio")
-    # Define initial evaporator heat source
-    sourceFluid = "Air"
-    sourceIn_m = 5.5  # [kg/s]
-    sourceIn_T = 744.  # [K], from Table 9
-    sourceIn_p = mc.bar2Pa(0.37)  # [Pa], from cp=1085.4
-    # Create FlowState of initial heat source
-    sourceIn = mc.FlowState(sourceFluid, -1, sourceIn_m, CP.PT_INPUTS,
-                            sourceIn_p, sourceIn_T)
-    # Define geometries for evaporator channels
-    evapGeomPlateWf = mc.GeomHxPlateSmooth(b=0.024375)
-    evapGeomPlateSf = mc.GeomHxPlateFinOffset(
-        s=0.008, h=0.03, t=0.0002, l=0.02)
-    # Define plate material & temp to take properties at
-    plateMaterial = mc.stainlessSteel_316(T=293.15)
-    # Define ambient conditions
-    altitude = mc.ft2m(20000)
-    mach = 0.49  # Mach number of aircraft
-    stillAmbient = mc.isa(altitude)  # ISA conditions at 20,000 ft
-    ambient_T = stillAmbient.T() / mc.TTotalRatio(gamma=1.4, M=mach)
-    ambient_p = stillAmbient.p() / mc.pTotalRatio(gamma=1.4, M=mach)
-    ambient = mc.FlowState('air', -1, 0, CP.PT_INPUTS, ambient_p, ambient_T)
-    # Create evaporator object
-    evap = mc.HxPlate(
-        flowSense='counterflow',
-        NPlate=17,  # total number of plates
-        RfWf=0,  # no fouling in working fluid channels
-        RfSf=0,  # no fouling in secondary fluid channels
-        plate=plateMaterial,
-        tPlate=0.2e-3,  # plate thickness [m]
-        geomPlateWf=evapGeomPlateWf,
-        geomPlateSf=evapGeomPlateSf,
-        L=0.2,  # heat trans. length [m]
-        W=0.44,  # heat trans. width [m]
-        effThermal=1.0,  # thermal efficiency
-        ambient=ambient,  # FlowState of ambient
-    )
[…]
