### 2018-07-31 ee0f17a
Update setup.py description
