### 2018-07-31 979eec1
Update README.rst
