### 2018-08-02 #710 @arfon
@momargoh - your paper is now accepted and your DOI is https://doi.org/10.21105/joss.00710 :zap: :rocket: :boom:
