### 2018-08-02 #710 @momargoh
@katyhuff @arfon

Thanks so much, however my paper incorrectly says the software has the old CC-BY licence, which was updated to Apache-2.0. I had updated the github license.txt file as well as paper.json to reflect this. Is it possible to rectify this?
